+++ src/main/java/com/atlauncher/data/Instance.java
@@ -587,8 +587,18 @@ public class Instance extends MinecraftVersion {
         progressDialog.setLabel(GetText.tr("Organising Libraries"));
 
         // extract natives to a temp dir
+        boolean useSystemGlfw = Optional.ofNullable(launcher.useSystemGlfw).orElse(App.settings.useSystemGlfw);
+        boolean useSystemOpenAl = Optional.ofNullable(launcher.useSystemOpenAl).orElse(App.settings.useSystemOpenAl);
+        System.out.println(useSystemGlfw);
+        System.out.println(useSystemOpenAl);
         this.libraries.stream().filter(Library::shouldInstall).forEach(library -> {
             if (library.hasNativeForOS()) {
+                if ((library.name.contains("glfw") && useSystemGlfw)
+                        || (library.name.contains("openal") && useSystemOpenAl)) {
+                    System.out.println(library.name);
+                    return;
+                }
+
                 Path nativePath = FileSystem.LIBRARIES.resolve(library.getNativeDownloadForOS().path);
 
                 ArchiveUtils.extract(nativePath, nativesTempDir, name -> {