+++ src/main/java/com/atlauncher/data/Settings.java
@@ -92,6 +92,8 @@ public class Settings {
     public boolean ignoreJavaOnInstanceLaunch = false;
     public boolean useJavaProvidedByMinecraft = true;
     public boolean disableLegacyLaunching = false;
+    public boolean useSystemGlfw = false;
+    public boolean useSystemOpenAl = false;
 
     // Network
     public int concurrentConnections = 8;