+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 ### New Features
 - Add in a new packs browser containing all platforms packs in one place
 - Run `-initSettings` with the server jar to generate eula.txt and server.properties [#523]
+- Add settings to use system install for GLFW/OpenAL [#524]
 
 ### Fixes
 - Issue with PermSize option being used on custom Java paths in some cases [#520]